vpn: add support for UDP broadcast
